Bug fix in PowerBI file
For the top 5 products, it can be seen that the first value was called "Blank", this is because for this graph the product code was used from the products table and in the transactions table there are 57 more products that are not included in the products table (error in the database I suppose). So to fix this bug, the product code must be taken as a reference from the transactions table.